+++ products/DesignerApp.browser.js
@@ -15,7 +15,7 @@ class DesignerCommander extends AbstractCommander {
     jQuery("#compileResultsDiv").val(this.program.compile())
   }
   visualizeCommand() {
-    jQuery("#explainResultsDiv").html(this._toIceCubes(this.program))
+    jQuery("#explainResultsDiv").html(this._toIceTray(this.program))
   }
   inferPrefixGrammarCommand() {
     this._app.setGrammarCode(new jtree.UnknownGrammarProgram(this._app.getCodeValue()).inferGrammarFileForAKeywordLanguage("inferredLanguage"))
@@ -61,25 +61,26 @@ class DesignerCommander extends AbstractCommander {
       saveAs(content, fileName)
     })
   }
-  _toIceCubes(program) {
+  _toIceTray(program) {
     const columns = program.getProgramWidth()
     const cellTypes = new jtree.TreeNode(program.getInPlaceCellTypeTreeWithNodeConstructorNames())
     const rootCellTypes = new jtree.TreeNode(program.getInPlacePreludeCellTypeTreeWithNodeConstructorNames())
     const table = program
       .getProgramAsCells()
       .map((line, lineIndex) => {
-        let rows = ""
+        const nodeType = cellTypes.nodeAt(lineIndex).getWord(0)
+        let cells = `<td class="iceTrayNodeType">${nodeType}</td>` // todo: add ancestry
         for (let cellIndex = 0; cellIndex < columns; cellIndex++) {
           const cell = line[cellIndex]
-          if (!cell) rows += `<td>&nbsp;</td>`
+          if (!cell) cells += `<td>&nbsp;</td>`
           else {
             const cellType = cellTypes.nodeAt(lineIndex).getWord(cellIndex + 1)
             const rootCellType = rootCellTypes.nodeAt(lineIndex).getWord(cellIndex + 1)
-            const nodeType = cellTypes.nodeAt(lineIndex).getWord(0)
-            rows += `<td title="cellType:${cellType} rootCellType:${rootCellType} nodeType:${nodeType}">${cell.getWord()}</td>`
+            const cellTypeDivs = [cellType, rootCellType] // todo: add full ancestry
+            cells += `<td><span class="cellTypeSpan">${cellTypeDivs.join(" ")}</span>${cell.getWord()}</td>`
           }
         }
-        return `<tr>${rows}</tr>`
+        return `<tr>${cells}</tr>`
       })
       .join("\n")
     return `<table class="iceCubes">${table}</table>`
@@ -88,7 +89,7 @@ class DesignerCommander extends AbstractCommander {
 class DesignerApp extends AbstractTreeComponentRootNode {
   constructor() {
     super(...arguments)
-    this.languages = "newlang hakon stump dumbdown arrow dug iris fire wwt swarm project stamp grammar config jibberish numbers poop".split(" ")
+    this.languages = "newlang hakon stump dumbdown arrow dug iris fire chuck wwt swarm project stamp grammar config jibberish numbers poop".split(" ")
     this._localStorageKeys = {
       grammarConsole: "grammarConsole",
       codeConsole: "codeConsole"
@@ -312,8 +313,30 @@ td
 .iceCubes
  tr,td
   margin 0
-  box-shadow rgba(1,1,1,.4) 1px 1px 1px
   overflow scroll
+  border 0
+ td
+  box-shadow rgba(1,1,1,.1) 1px 1px 1px
+  position relative
+  padding 10px 3px 2px 2px
+  .cellTypeSpan
+   position absolute
+   white-space nowrap
+   left 0
+   top 0
+   font-size 8px
+   color rgba(1,1,1,.2)
+ .iceTrayNodeType
+  box-shadow none
+  font-size 8px
+  color rgba(1,1,1,.2)
+ tr
+  &:hover
+   td
+    .iceTrayNodeType
+     color rgba(1,1,1,.5)
+    .cellTypeSpan
+     color rgba(1,1,1,.5)
 code
  white-space pre
 pre